+++ internal/server/hostile2.go
@@ -32,8 +32,9 @@ var (
 	entityStray    = entityID("stray")
 	entityWitch    = entityID("witch")
 
-	entityPearlProj  = entityID("ender_pearl")   // thrown ender pearl
-	entitySplashProj = entityID("splash_potion") // witch's splash potion
+	entityPearlProj  = entityID("ender_pearl")      // thrown ender pearl
+	entitySplashProj = entityID("splash_potion")    // witch's splash potion
+	entityLingerProj = entityID("lingering_potion") // a thrown lingering potion: its own entity type
 
 	itemSlimeball  = itemByName["slime_ball"]
 	itemEnderPearl = itemByName["ender_pearl"]